change rating from int to float

+++ app/src/main/java/com/nabil/rateme/model/Movie.kt
@@ -9,5 +9,5 @@ data class Movie(
     @PrimaryKey(autoGenerate = true) var id: Long? = null,
     var name: String,
     var image: Int,
-    var rating: Int
+    var rating: Float
 )
